JBIDE-16136 Reorganize jst.test plugins
