+++ cdi/tests/org.jboss.tools.cdi.extension.core.test/META-INF/MANIFEST.MF
@@ -12,7 +12,7 @@ Require-Bundle: org.eclipse.jdt.core,
  org.jboss.tools.tests,
  org.eclipse.core.runtime;bundle-version="3.7.0",
  org.jboss.tools.cdi.core,
- org.jboss.tools.jst.jsp.base.test,
+ org.jboss.tools.jst.web.ui.test,
  org.jboss.tools.cdi.core.test,
  org.jboss.tools.cdi.extension.core,
  org.jboss.tools.common,
@@ -21,7 +21,6 @@ Require-Bundle: org.eclipse.jdt.core,
  org.eclipse.jface.text;bundle-version="3.7.0",
  org.jboss.tools.common.el.core,
  org.jboss.tools.common.text.ext,
- org.jboss.tools.jst.text.ext.test,
  org.eclipse.wst.validation;bundle-version="1.2.300",
  org.jboss.tools.jst.web.ui
 Export-Package: org.jboss.tools.cdi.extension.core.test
