+++ cdi/tests/org.jboss.tools.cdi.seam.core.test/META-INF/MANIFEST.MF
@@ -13,7 +13,7 @@ Require-Bundle: org.jboss.tools.cdi.seam.core,
  org.jboss.tools.tests,
  org.eclipse.core.runtime;bundle-version="3.7.0",
  org.jboss.tools.cdi.core,
- org.jboss.tools.jst.jsp.base.test,
+ org.jboss.tools.jst.web.ui.test,
  org.jboss.tools.cdi.core.test,
  org.jboss.tools.common,
  org.eclipse.ui;bundle-version="3.7.0",
@@ -21,7 +21,6 @@ Require-Bundle: org.jboss.tools.cdi.seam.core,
  org.eclipse.jface.text;bundle-version="3.7.0",
  org.jboss.tools.common.el.core,
  org.jboss.tools.common.text.ext,
- org.jboss.tools.jst.text.ext.test,
  org.jboss.tools.cdi.seam.solder.core.test,
  org.eclipse.wst.validation;bundle-version="1.2.300",
  org.jboss.tools.jst.web.ui
