+++ cdi/tests/org.jboss.tools.cdi.seam.core.test/src/org/jboss/tools/cdi/seam/core/test/international/CACdiAddInfoELMessagesTest.java
@@ -17,8 +17,8 @@ import java.util.Set;
 import org.eclipse.jface.text.contentassist.ICompletionProposal;
 import org.jboss.tools.cdi.seam.core.test.SeamCoreTest;
 import org.jboss.tools.common.base.test.contentassist.CATestUtil;
+import org.jboss.tools.jst.web.ui.base.test.ca.ContentAssistantTestCase;
 import org.jboss.tools.jst.web.ui.internal.editor.contentassist.AutoELContentAssistantProposal;
-import org.jboss.tools.jst.jsp.test.ca.ContentAssistantTestCase;
 
 /**
  * The JUnit test case for JBIDE-9910 issue
