+++ cdi/tests/org.jboss.tools.cdi.seam.faces.core.test/META-INF/MANIFEST.MF
@@ -12,7 +12,7 @@ Require-Bundle: org.jboss.tools.cdi.seam.faces.core,
  org.jboss.tools.tests,
  org.eclipse.core.runtime;bundle-version="3.7.0",
  org.jboss.tools.cdi.core,
- org.jboss.tools.jst.jsp.base.test,
+ org.jboss.tools.jst.web.ui.test,
  org.eclipse.wst.validation,
  org.jboss.tools.cdi.core.test,
  org.jboss.tools.common.validation,
