+++ cdi/tests/org.jboss.tools.cdi.ui.test/META-INF/MANIFEST.MF
@@ -23,7 +23,7 @@ Require-Bundle: org.junit,
  org.eclipse.jdt.core,
  org.eclipse.jdt.ui,
  org.eclipse.ui.ide,
- org.jboss.tools.jst.jsp.base.test,
+ org.jboss.tools.jst.web.ui.test,
  org.eclipse.ltk.core.refactoring,
  org.eclipse.search,
  org.jboss.tools.common.el.core,
