+++ jsf/tests/org.jboss.tools.jsf.text.ext.test/META-INF/MANIFEST.MF
@@ -21,7 +21,7 @@ Require-Bundle: org.junit,
  org.jboss.tools.jsf.text.ext.richfaces,
  org.jboss.tools.jsf.ui,
  org.eclipse.jst.standard.schemas,
- org.jboss.tools.jst.text.ext.test;bundle-version="3.3.0",
+ org.jboss.tools.jst.web.ui.test,
  org.eclipse.jdt.ui;bundle-version="3.7.0",
  org.jboss.tools.jsf;bundle-version="3.3.0"
 Bundle-ActivationPolicy: lazy
