+++ jsf/tests/org.jboss.tools.jsf.text.ext.test/src/org/jboss/tools/jsf/text/ext/test/JSF2MessagesOpenOnTest.java
@@ -19,9 +19,9 @@ import org.eclipse.core.resources.ResourcesPlugin;
 import org.eclipse.ui.PlatformUI;
 import org.jboss.tools.jst.web.ui.internal.text.ext.hyperlink.ELHyperlink;
 import org.jboss.tools.jst.web.ui.internal.text.ext.hyperlink.ELHyperlinkDetector;
-import org.jboss.tools.jst.text.ext.test.HyperlinkTestUtil;
-import org.jboss.tools.jst.text.ext.test.HyperlinkTestUtil.TestRegion;
-import org.jboss.tools.jst.text.ext.test.HyperlinkTestUtil.TestHyperlink;
+import org.jboss.tools.jst.web.ui.openon.test.HyperlinkTestUtil;
+import org.jboss.tools.jst.web.ui.openon.test.HyperlinkTestUtil.TestHyperlink;
+import org.jboss.tools.jst.web.ui.openon.test.HyperlinkTestUtil.TestRegion;
 
 /**
  *
