+++ jsf/tests/org.jboss.tools.jsf.ui.test/META-INF/MANIFEST.MF
@@ -34,7 +34,6 @@ Require-Bundle: org.junit,
  org.eclipse.wst.validation,
  org.eclipse.wst.html.core,
  org.eclipse.wst.html.ui,
- org.jboss.tools.jst.jsp.base.test,
  org.jboss.tools.common.el.ui,
  org.eclipse.jdt.ui,
  org.jboss.tools.common.ui,
