+++ jsf/tests/org.jboss.tools.jsf.ui.test/src/org/jboss/tools/jsf/jsp/ca/test/CAForELinStyleTest.java
@@ -1,9 +1,9 @@
 package org.jboss.tools.jsf.jsp.ca.test;
 
-import org.jboss.tools.jst.jsp.test.ca.ContentAssistantTestCase;
+import org.jboss.tools.jst.web.ui.base.test.ca.ContentAssistantTestCase;
 import org.jboss.tools.test.util.TestProjectProvider;
 
-public class CAForELinStyleTest extends ContentAssistantTestCase{
+public class CAForELinStyleTest extends ContentAssistantTestCase {
 	TestProjectProvider provider = null;
 	boolean makeCopy = false;
 	private static final String PROJECT_NAME = "CAForELinStyleTest";
