+++ jsf/tests/org.jboss.tools.jsf.ui.test/src/org/jboss/tools/jsf/jsp/ca/test/CAForIDTest.java
@@ -1,7 +1,7 @@
 package org.jboss.tools.jsf.jsp.ca.test;
 
 import org.eclipse.core.resources.IncrementalProjectBuilder;
-import org.jboss.tools.jst.jsp.test.ca.ContentAssistantTestCase;
+import org.jboss.tools.jst.web.ui.base.test.ca.ContentAssistantTestCase;
 import org.jboss.tools.test.util.TestProjectProvider;
 
 public class CAForIDTest extends ContentAssistantTestCase{
