+++ seam/tests/org.jboss.tools.seam.ui.test/META-INF/MANIFEST.MF
@@ -38,11 +38,10 @@ Require-Bundle: org.junit,
  org.eclipse.jst.jee,
  org.eclipse.jst.jee.ejb,
  org.jboss.tools.seam.base.test,
- org.jboss.tools.jst.jsp.base.test,
+ org.jboss.tools.jst.web.ui.test,
  org.jboss.tools.tests,
  org.jboss.tools.jst.web,
  org.eclipse.jst.jee.web,
- org.jboss.tools.jst.text.ext.test,
  org.jboss.tools.jst.web.kb.test,
  org.eclipse.jst.servlet.ui;bundle-version="1.1.500",
  org.hibernate.eclipse.console,
